Tidied up the IEC type classes INT and UINT.
